working more on new design
