+++ src/components/pdfs/output.css
@@ -183,3 +183,49 @@ span {
   display: flex;
   flex-direction: column;
 }
+
+
+/* ---------- new styles --------------- */
+
+.wide-box-1 {
+  width: 800px;
+  height: 130px;
+  border: 3px solid black;
+  border-radius: 5px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  padding-left: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 8px 0;
+}
+
+.row {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+}
+
+.dataa {
+  font-weight: bold;
+  width: 33%;
+}
+
+.heading {
+  padding-right: 5px;
+}
+
+.wide-box-2 {
+  width: 800px;
+  height: 85px;
+  border: 3px solid black;
+  border-radius: 5px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  padding-left: 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 8px 0;
+}
